Fixed a small bug with compress/flate/Inflater. It incorrectly checked the size of chunk with uncompressed data (00 is terms of DEFLATE).
APPROVED=rsc
DELTA=15  (14 added, 0 deleted, 1 changed)
OCL=32105
CL=32238

+++ src/pkg/compress/flate/inflate.go
@@ -542,7 +542,7 @@ func (f *inflater) dataBlock() os.Error {
 	}
 	n := int(f.buf[0]) | int(f.buf[1])<<8;
 	nn := int(f.buf[2]) | int(f.buf[3])<<8;
-	if nn != ^n {
+	if uint16(nn) != uint16(^n) {
 		return CorruptInputError(f.roffset);
 	}
